moved sheet view to a separate file
